Move reducer
