+++ src/reducers.js
@@ -1,10 +1,9 @@
 import { combineReducers } from 'redux';
-import activeMetaContainer from './components/activeMetaContainer/reducer';
-// import visibilityFilter from './visibilityFilter'
+import homeReducer from './routes/home/reducer';
 
 const reducers = combineReducers({
-  activeMetaContainer
-  // visibilityFilter
+  home: homeReducer
+  // add more reducers here
 });
 
 export default reducers;
